move tests to a single test file
